updated location redirect overview (chua on phan redirect check login)
