+++ src/page/client/pageClient/personal/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import { MdNavigateNext } from 'react-icons/md';
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
@@ -8,6 +8,10 @@ import ListLikeSong from "./page/listLikeSong"
 import ListArtist from "./page/listArtist";
 import Playlist from "./page/playlistSong";
 import Upload from "./page/upload"
+import { useSelector } from 'react-redux';
+import { formStateUser } from 'redux/user/stateUser';
+import Notification from 'page/notificationModal/NotificationModal';
+import { useHistory } from 'react-router-dom';
 interface IndexPersonal<T> {
 
 }
@@ -23,7 +27,17 @@ export interface PropsPersonal {
 }
 
 const IndexPersonal: React.FC<IndexPersonal<any>> = ({ ...props }) => {
+    const userState = useSelector<{ user: any }>(state => state.user) as formStateUser;
+    const [isLogged, setIsLogged] = useState(false);
     const [page, setpage] = useState<string>("");
+    const history = useHistory();
+    if(userState.token === ''){
+        history.push('/', {isLogged: false});
+        return <></>;
+    }
+
+    console.log(history)
+
     const renderPage = <T extends string>(page?: T) => {
         switch (page) {
             case AllPage.OverView: return <OverView navigatePage={navigatePage} />
@@ -38,8 +52,15 @@ const IndexPersonal: React.FC<IndexPersonal<any>> = ({ ...props }) => {
         if (!page) return
         setpage(page)
     }
+
+    const handleLogged = () => {
+        setIsLogged(false);
+    }
+
     return (
         <>
+        {isLogged && <Notification handleLogged={handleLogged} />}
+
             <div className="Personal">
                 <Tabed navigatePage={navigatePage} />
                 <div className="overview">
