+++ src/page/partials/client/sidebar.tsx
@@ -1,44 +1,28 @@
 
-import React, { useEffect, useState } from 'react'
+import React from 'react'
 import { BsMusicNoteBeamed,BsListUl } from 'react-icons/bs';
 import { FaBlogger, FaChartPie } from 'react-icons/fa';
 import { RiFolderMusicFill } from 'react-icons/ri';
-import { Link, Redirect, RouteChildrenProps, withRouter, Route, useHistory } from "react-router-dom";
+import { Link, useLocation } from "react-router-dom";
 import { BsPlusCircle } from 'react-icons/bs';
 import { BiPlayCircle,BiTimeFive } from 'react-icons/bi';
 import { AiFillStar,AiOutlineHeart } from 'react-icons/ai';
-import { useSelector } from 'react-redux';
-import { formStateUser } from 'redux/user/stateUser';
 import Popup from '@titaui/reactjs-popup';
-import Notification from 'page/notificationModal/NotificationModal';
+
 interface Sidebar<T> {
 
 }
 
 const Sidebar: React.FC<Sidebar<any>> = ({ ...props }) => {
-  const userState = useSelector<{ user: any }>(state => state.user) as formStateUser;
-  const [isLogged, setIsLogged] = useState(false);
-  const history = useHistory();
+  const location = useLocation();
 
-  const clickedOverview = (path: string) => {
-    console.log(path)
-    if(userState.token){
-      return history.push(path);
-    }
-
-    setIsLogged(true)
-  }
-  const handleLogged = () => {
-    setIsLogged(false);
-  }
-
+  console.log("this is sidebar location: ", location)
   return (
     <>
       <div className="sidebar">
-        {isLogged && <Notification handleLogged={handleLogged} />}
         <h5><Link to="/">MUSIC GAME</Link></h5>
         <ul>
-        <Link to="#" onClick={()=>clickedOverview('/overview')}><li><BsMusicNoteBeamed className="icon" />Cá nhân</li></Link>
+        <Link to={{pathname: '/overview', state: {page: "OverView"}}}><li><BsMusicNoteBeamed className="icon" />Cá nhân</li></Link>
         {/* <Link to="/overview"><li><BsMusicNoteBeamed className="icon" />Cá nhân</li></Link> */}
         <Link to="/"><li><RiFolderMusicFill className="icon" />Khám phá</li></Link>
         <Link to="/chart"><li><FaChartPie className="icon" />Music chart</li></Link>
@@ -52,9 +36,9 @@ const Sidebar: React.FC<Sidebar<any>> = ({ ...props }) => {
           </ul>
           <ul>
             <h6>Thư viện</h6>
-            <Link to="#" onClick={()=>clickedOverview('/favorite')}><li><AiOutlineHeart className="icon" />Yêu thích</li></Link>
-            <Link to="/music"><li><BsMusicNoteBeamed className="icon" />Bài hát</li></Link>
-            <Link to="#" onClick={()=>clickedOverview('/playlist')}><li><BiPlayCircle className="icon" />Playlist</li></Link>
+            <Link to={{pathname: '/overview', state: {page: "OverView", lastLocation: location.pathname}}}><li><AiOutlineHeart className="icon" />Yêu thích</li></Link>
+            <Link to={{pathname: '/overview', state: {page: "Music", lastLocation: location.pathname}}}><li><BsMusicNoteBeamed className="icon" />Bài hát</li></Link>
+            <Link to={{pathname: '/overview', state: {page: "Playlist", lastLocation: location.pathname}}}><li><BiPlayCircle className="icon" />Playlist</li></Link>
             <Link to="/recently"><li><BiTimeFive className="icon" />Gần đây</li></Link>
           </ul>
           <ul>
